ChromaDB: Add connection to a remote server
ChromaDB can be run in a remote context using docker. However, there
is no inbuilt authentication solution yet according to the documentation.

Therefore, this method is only useful if a user want to connect to an
unauthenticated ChromaDB instance. Only use this on a local network
or with mesh solutions such as Tailscale/Zerotier.

Signed-off-by: kingbri <bdashore3@proton.me>